Remove puppeteer-driven og generator
