+++ utils/ogImage.ts
@@ -1,34 +1,12 @@
 import { Atoms } from "@/designSystem"
 import Canvas from "canvas"
 import widont from "./widont"
+import siteConfig from "../siteconfig.json"
 
 const { createCanvas, registerFont } = Canvas
 
-const namesToTry = [
-  "Söhne",
-  "Söhne Breit",
-  "Söhne Breit Fett",
-  "Soehne",
-  "Soehne Breit",
-  "Soehne Breit Fett",
-  "SöhneBreit",
-  "SöhneBreit-Fett",
-  "SoehneBreit",
-  "SoehneBreit-Fett",
-  "Sohne",
-  "Sohne Breit",
-  "Sohne Breit Fett",
-  "SohneBreit",
-  "SohneBreit-Fett",
-  "S\u00F6hne",
-  "S\u00F6hne Breit",
-  "S\u00F6hne Breit Fett",
-  "S\u00F6hneBreit",
-  "S\u00F6hneBreit-Fett",
-]
-
 const fontNames = {
-  soehne: namesToTry[19],
+  soehne: "Soehne",
   national: "National 2",
 }
 
@@ -64,32 +42,53 @@ function getLines(
   lines.push(currentLine)
   return lines
 }
+
 export default function ogImage(title: string): Buffer | null {
   if (!title) {
     return null
   }
 
   const displaySize = 80
+  const lineHeight = displaySize * 1
+  const authorSize = 36
   const [width, height] = [1200, 1200]
   const [halfWidth, halfHeight] = [width / 2, height / 2]
-  const textWidth = 1024
+  const textWidth = 1022
   const offset = 66
 
   const canvas = createCanvas(width, height)
   const ctx = canvas.getContext("2d")
 
+  // Render the background colour
   ctx.fillStyle = Atoms.colors.text
   ctx.fillRect(0, 0, width, height)
 
+  // Render the blog post title
   ctx.font = `bold ${displaySize}px ${fontNames.soehne}`
   ctx.textAlign = "left"
   ctx.textBaseline = "middle"
   ctx.fillStyle = Atoms.colors.wash
 
   const lines = getLines(ctx, widont(String(title)), textWidth)
-  const textHeight = lines.length * displaySize
+  const textHeight = lines.length * lineHeight
 
-  ctx.fillText(lines.join("\n"), offset, halfHeight - textHeight / 2)
+  lines.map((line, i) => {
+    const y = halfHeight - textHeight / 2 + lineHeight * i + lineHeight / 3
+    ctx.fillText(line, offset, y)
+  })
+
+  // Render the site name
+  if (title !== siteConfig.title) {
+    ctx.font = `${authorSize}px ${fontNames.national}`
+    ctx.fillStyle = Atoms.colors.meta
+
+    ctx.translate(width - offset, halfHeight)
+    ctx.rotate(-Math.PI / 2)
+    ctx.textAlign = "center"
+    ctx.fillText(siteConfig.title, 0, 0)
+    ctx.resetTransform()
+    ctx.restore()
+  }
 
   const buffer = canvas.toBuffer("image/png")
